Move golden fixtures under tests/golden/ per #140 convention
Align with tests/golden/fixtures/ layout established by #140.

- tests/fixtures/golden/*.parquet -> tests/golden/fixtures/
- tests/fixtures/generate_rolling_golden.py -> tests/golden/
- update GOLDEN_DIR, module path in docstring, and 3 test refs
