+++ tests/golden/fixtures/prc_to_high_21d.parquet
