+++ tests/golden/fixtures/turnover_126d.parquet
