+++ tests/golden/fixtures/zero_trades_126d.parquet
